Update DSA_P114. TONG LON NHAT CUA DAY CON LIEN TUC.cpp

+++ BAI THUC HANH SO 1 (Chủ đề Sinh kế tiếp, Quay lui - Nhánh cận, Giải thuật tham lam, Chia và trị, Quy hoạch động)/WEEK 2 (19 - 03 - 2023)/DSA_P114. TONG LON NHAT CUA DAY CON LIEN TUC.cpp
@@ -30,7 +30,7 @@ int main(){
         }
         long sum = 0;
         for (int i = 0; i < n; i++) {
-            sum += (arr[i]) * (i - left[i]) * (right[i] - i);
+            sum += (long long) (arr[i]) * (i - left[i]) * (right[i] - i);
         }
         cout<<sum<<endl;
     }
